+++ Src/libCZI/SingleChannelPyramidLevelTileAccessor.cpp
@@ -10,14 +10,20 @@
 using namespace libCZI;
 using namespace std;
 
-CSingleChannelPyramidLevelTileAccessor::CSingleChannelPyramidLevelTileAccessor(std::shared_ptr<ISubBlockRepository> sbBlkRepository)
+CSingleChannelPyramidLevelTileAccessor::CSingleChannelPyramidLevelTileAccessor(const std::shared_ptr<ISubBlockRepository>& sbBlkRepository)
     : CSingleChannelAccessorBase(sbBlkRepository)
 {
 }
 
 /*virtual*/std::shared_ptr<libCZI::IBitmapData> CSingleChannelPyramidLevelTileAccessor::Get(const libCZI::IntRect& roi, const libCZI::IDimCoordinate* planeCoordinate, const PyramidLayerInfo& pyramidInfo, const ISingleChannelPyramidLayerTileAccessor::Options* pOptions)
 {
-    if (pOptions == nullptr) { Options opt; opt.Clear(); return this->Get(roi, planeCoordinate, pyramidInfo, &opt); }
+    if (pOptions == nullptr)
+    {
+        Options opt;
+        opt.Clear();
+        return this->Get(roi, planeCoordinate, pyramidInfo, &opt);
+    }
+
     libCZI::PixelType pixelType;
     const bool b = this->TryGetPixelType(planeCoordinate, pixelType);
     if (b == false)
@@ -30,7 +36,13 @@ CSingleChannelPyramidLevelTileAccessor::CSingleChannelPyramidLevelTileAccessor(s
 
 /*virtual*/std::shared_ptr<libCZI::IBitmapData> CSingleChannelPyramidLevelTileAccessor::Get(libCZI::PixelType pixeltype, const libCZI::IntRect& roi, const libCZI::IDimCoordinate* planeCoordinate, const PyramidLayerInfo& pyramidInfo, const libCZI::ISingleChannelPyramidLayerTileAccessor::Options* pOptions)
 {
-    if (pOptions == nullptr) { Options opt; opt.Clear(); return this->Get(pixeltype, roi, planeCoordinate, pyramidInfo, &opt); }
+    if (pOptions == nullptr)
+    {
+        Options opt;
+        opt.Clear();
+        return this->Get(pixeltype, roi, planeCoordinate, pyramidInfo, &opt);
+    }
+
     const int sizeOfPixel = CalcSizeOfPixelOnLayer0(pyramidInfo);
     const IntSize sizeOfBitmap{ static_cast<std::uint32_t>(roi.w / sizeOfPixel),static_cast<std::uint32_t>(roi.h / sizeOfPixel) };
     if (sizeOfBitmap.w == 0 || sizeOfBitmap.h == 0)
@@ -46,7 +58,14 @@ CSingleChannelPyramidLevelTileAccessor::CSingleChannelPyramidLevelTileAccessor(s
 
 /*virtual*/void CSingleChannelPyramidLevelTileAccessor::Get(libCZI::IBitmapData* pDest, int xPos, int yPos, const libCZI::IDimCoordinate* planeCoordinate, const PyramidLayerInfo& pyramidInfo, const Options* pOptions)
 {
-    if (pOptions == nullptr) { Options opt; opt.Clear(); this->Get(pDest, xPos, yPos, planeCoordinate, pyramidInfo, &opt); return; }
+    if (pOptions == nullptr)
+    {
+        Options opt;
+        opt.Clear();
+        this->Get(pDest, xPos, yPos, planeCoordinate, pyramidInfo, &opt);
+        return;
+    }
+
     const int sizeOfPixel = CalcSizeOfPixelOnLayer0(pyramidInfo);
     this->InternalGet(pDest, xPos, yPos, sizeOfPixel, planeCoordinate, pyramidInfo, *pOptions);
 }
@@ -74,7 +93,7 @@ void CSingleChannelPyramidLevelTileAccessor::InternalGet(libCZI::IBitmapData* pD
         });
 }
 
-void CSingleChannelPyramidLevelTileAccessor::ComposeTiles(libCZI::IBitmapData* bm, int xPos, int yPos, int sizeOfPixel, int bitmapCnt, const Options& options, std::function<SbInfo(int)> getSbInfo)
+void CSingleChannelPyramidLevelTileAccessor::ComposeTiles(libCZI::IBitmapData* bm, int xPos, int yPos, int sizeOfPixel, int bitmapCnt, const Options& options, const std::function<SbInfo(int)>& getSbInfo)
 {
     Compositors::ComposeSingleTileOptions composeOptions; composeOptions.Clear();
     composeOptions.drawTileBorder = options.drawTileBorder;
@@ -92,13 +111,12 @@ void CSingleChannelPyramidLevelTileAccessor::ComposeTiles(libCZI::IBitmapData* b
                 return true;
             }
 
-    return false;
+            return false;
         },
         bm,
-            0,
-            0,
-            &composeOptions);
-
+        0,
+        0,
+        &composeOptions);
 }
 
 libCZI::IntRect CSingleChannelPyramidLevelTileAccessor::CalcDestinationRectFromSourceRect(const libCZI::IntRect& roi, const PyramidLayerInfo& pyramidInfo)
@@ -226,16 +244,16 @@ void CSingleChannelPyramidLevelTileAccessor::GetAllSubBlocks(const libCZI::IntRe
                 }
             }
 
-    if (Utilities::DoIntersect(roi, info.logicalRect))
-    {
-        SbInfo sbinfo;
-        sbinfo.logicalRect = info.logicalRect;
-        sbinfo.physicalSize = info.physicalSize;
-        sbinfo.mIndex = info.mIndex;
-        sbinfo.index = idx;
-        appender(sbinfo);
-    }
+            if (Utilities::DoIntersect(roi, info.logicalRect))
+            {
+                SbInfo sbinfo;
+                sbinfo.logicalRect = info.logicalRect;
+                sbinfo.physicalSize = info.physicalSize;
+                sbinfo.mIndex = info.mIndex;
+                sbinfo.index = idx;
+                appender(sbinfo);
+            }
 
-    return true;
+            return true;
         });
 }