+++ Src/libCZI/SingleChannelTileAccessor.cpp
@@ -7,13 +7,12 @@
 #include "utilities.h"
 #include "SingleChannelTileCompositor.h"
 #include "Site.h"
-#include <iterator> 
 #include "bitmapData.h"
 
 using namespace libCZI;
 using namespace std;
 
-CSingleChannelTileAccessor::CSingleChannelTileAccessor(std::shared_ptr<ISubBlockRepository> sbBlkRepository)
+CSingleChannelTileAccessor::CSingleChannelTileAccessor(const std::shared_ptr<ISubBlockRepository>& sbBlkRepository)
     : CSingleChannelAccessorBase(sbBlkRepository)
 {
 }
@@ -45,26 +44,66 @@ CSingleChannelTileAccessor::CSingleChannelTileAccessor(std::shared_ptr<ISubBlock
 
 void CSingleChannelTileAccessor::ComposeTiles(libCZI::IBitmapData* pBm, int xPos, int yPos, const std::vector<IndexAndM>& subBlocksSet, const ISingleChannelTileAccessor::Options& options)
 {
-    Compositors::ComposeSingleTileOptions composeOptions; composeOptions.Clear();
+    Compositors::ComposeSingleTileOptions composeOptions;
+    composeOptions.Clear();
     composeOptions.drawTileBorder = options.drawTileBorder;
-    Compositors::ComposeSingleChannelTiles(
-        [&](int index, std::shared_ptr<libCZI::IBitmapData>& spBm, int& xPosTile, int& yPosTile)->bool
-        {
-            if (index < static_cast<int>(subBlocksSet.size()))
-            {
-                const auto sb = this->sbBlkRepository->ReadSubBlock(subBlocksSet[index].index);
-                spBm = sb->CreateBitmap();
-                xPosTile = sb->GetSubBlockInfo().logicalRect.x;
-                yPosTile = sb->GetSubBlockInfo().logicalRect.y;
-                return true;
-            }
 
-    return false;
-        },
-        pBm,
+    if (options.useVisibilityCheckOptimization)
+    {
+        // Try to reduce the number of subblocks to be rendered by doing a visibility check, and only rendering those which are visible.
+        // We report the subblocks in the order as they are given in the vector 'subBlocksSet', the lambda will be called with the
+        // argument 'index' counting down from subBlocksSet.size()-1 to 0. The subblock index we report for 'index=0' is the first one
+        // to be rendered, and 'index=subBlocksSet.size()-1' is the last one to be rendered (on top of all the others).
+        // We get a vector with the indices of the subblocks to be rendered, and then render them in the order as given in this vector 
+        // (index here means - the number as passed to the lambda).
+        const auto indices_of_visible_tiles = this->CheckForVisibility(
+            { xPos, yPos, static_cast<int>(pBm->GetWidth()), static_cast<int>(pBm->GetHeight()) },
+            static_cast<int>(subBlocksSet.size()),
+            [&](int index)->int
+            {
+                return subBlocksSet[index].index;
+            });
+
+        Compositors::ComposeSingleChannelTiles(
+            [&](int index, std::shared_ptr<libCZI::IBitmapData>& spBm, int& xPosTile, int& yPosTile)->bool
+            {
+                if (index < static_cast<int>(indices_of_visible_tiles.size()))
+                {
+                    const auto sb = this->sbBlkRepository->ReadSubBlock(subBlocksSet[indices_of_visible_tiles[index]].index);
+                    spBm = sb->CreateBitmap();
+                    xPosTile = sb->GetSubBlockInfo().logicalRect.x;
+                    yPosTile = sb->GetSubBlockInfo().logicalRect.y;
+                    return true;
+                }
+
+                return false;
+            },
+            pBm,
             xPos,
             yPos,
             &composeOptions);
+    }
+    else
+    {
+        Compositors::ComposeSingleChannelTiles(
+            [&](int index, std::shared_ptr<libCZI::IBitmapData>& spBm, int& xPosTile, int& yPosTile)->bool
+            {
+                if (index < static_cast<int>(subBlocksSet.size()))
+                {
+                    const auto sb = this->sbBlkRepository->ReadSubBlock(subBlocksSet[index].index);
+                    spBm = sb->CreateBitmap();
+                    xPosTile = sb->GetSubBlockInfo().logicalRect.x;
+                    yPosTile = sb->GetSubBlockInfo().logicalRect.y;
+                    return true;
+                }
+
+                return false;
+            },
+            pBm,
+            xPos,
+            yPos,
+            &composeOptions);
+    }
 }
 
 void CSingleChannelTileAccessor::InternalGet(int xPos, int yPos, libCZI::IBitmapData* pBm, const IDimCoordinate* planeCoordinate, const ISingleChannelTileAccessor::Options* pOptions)
@@ -79,13 +118,13 @@ void CSingleChannelTileAccessor::InternalGet(int xPos, int yPos, libCZI::IBitmap
     this->CheckPlaneCoordinates(planeCoordinate);
     Clear(pBm, pOptions->backGroundColor);
     const IntSize sizeBm = pBm->GetSize();
-    IntRect roi{ xPos,yPos,static_cast<int>(sizeBm.w),static_cast<int>(sizeBm.h) };
+    const IntRect roi{ xPos,yPos,static_cast<int>(sizeBm.w),static_cast<int>(sizeBm.h) };
     const std::vector<IndexAndM> subBlocksSet = this->GetSubBlocksSubset(roi, planeCoordinate, pOptions->sortByM);
 
     this->ComposeTiles(pBm, xPos, yPos, subBlocksSet, *pOptions);
 }
 
-std::vector<CSingleChannelTileAccessor::IndexAndM> CSingleChannelTileAccessor::GetSubBlocksSubset(const IntRect& roi, const IDimCoordinate* planeCoordinate, bool sortByM /*,libCZI::PixelType* pPixelTypeOfFirstFoundSubBlock=nullptr*/)
+std::vector<CSingleChannelTileAccessor::IndexAndM> CSingleChannelTileAccessor::GetSubBlocksSubset(const IntRect& roi, const IDimCoordinate* planeCoordinate, bool sortByM)
 {
     // ok... for a first tentative, experimental and quick-n-dirty implementation, simply
     // get all subblocks by enumerating all
@@ -106,7 +145,7 @@ std::vector<CSingleChannelTileAccessor::IndexAndM> CSingleChannelTileAccessor::G
     return subBlocksSet;
 }
 
-void CSingleChannelTileAccessor::GetAllSubBlocks(const IntRect& roi, const IDimCoordinate* planeCoordinate, std::function<void(int index, int mIndex)> appender/*, libCZI::PixelType* pPixelTypeOfFirstFoundSubBlock*/)
+void CSingleChannelTileAccessor::GetAllSubBlocks(const IntRect& roi, const IDimCoordinate* planeCoordinate, const std::function<void(int index, int mIndex)>& appender) const
 {
     this->sbBlkRepository->EnumSubset(planeCoordinate, nullptr, true,
         [&](int idx, const SubBlockInfo& info)->bool
@@ -116,7 +155,7 @@ void CSingleChannelTileAccessor::GetAllSubBlocks(const IntRect& roi, const IDimC
                 appender(idx, info.mIndex);
             }
 
-    return true;
+            return true;
         });
 }